Move ExpandingFormset tests to the component module
